# docs/swagger.yaml
basePath: /
definitions:
...
  database.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  database.SignUpDTO:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact: {}
  description: This is a sample Ticket-Booking API server for a Fiber app.
  title: Ticket-Booking API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to log in with their email and password.
      parameters:
      - description: User login details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/database.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Incorrect password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error generating token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: This endpoint allows users to register with their details.
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/database.SignUpDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User registration
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"